log: extract out log package
