+++ log/log.go
@@ -1,4 +1,4 @@
-package serve
+package log
 
 import (
 	"fmt"
